requirements.txt
Django==4.2.14
Pillow
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-gis
django-leaflet